Move MainServlet into its own file
